requirements.txt
rcssmin
rjsmin
minify-html
flask
requests
pillow